refs #236: split brave extension into core and spring boot
This was to remove warnings for building the Maven distributions in
regards to the optional spring dependencies. I feel it is easier to
understand and less likely to make mistakes if I just create separate
modules for them.
