+++ extensions/brave-extension/core/src/main/java/com/jashmore/sqs/extensions/brave/decorator/BraveMessageProcessingDecorator.java
