+++ extensions/brave-extension/core/src/test/java/com/jashmore/sqs/extensions/brave/decorator/BraveMessageProcessingDecoratorTest.java
