+++ extensions/brave-extension/spring-boot/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,1 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=com.jashmore.sqs.extensions.brave.spring.BraveMessageProcessingDecoratorConfiguration
