--- extensions/brave-message-processing-decorator/src/main/resources/META-INF/spring.factories
@@ -1,1 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=com.jashmore.sqs.extensions.brave.decorator.BraveMessageProcessingDecoratorConfiguration
